requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
confluent-kafka
stripe
python-dotenv
